+++ mercury/src/test/java/org/broadinstitute/gpinformatics/athena/boundary/orders/ProductOrderEjbTest.java
@@ -460,7 +460,7 @@ public class ProductOrderEjbTest {
 
         Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                 .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                        10000,1000000,conversionPdo,
+                        BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                         TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
         productOrderEjb.publishProductOrderToSAP(conversionPdo, messageCollection, true);
@@ -594,7 +594,7 @@ public class ProductOrderEjbTest {
         Mockito.when(mockSapClient.findMaterials(Mockito.anyString(), Mockito.anyString())).thenReturn(returnMaterials);
         Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                 .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                        10000,1000000,conversionPdo,
+                        BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                         TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
 
@@ -678,7 +678,7 @@ public class ProductOrderEjbTest {
         Mockito.when(mockSapClient.findMaterials(Mockito.anyString(), Mockito.anyString())).thenReturn(returnMaterials);
         Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                 .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                        10000,1000000,conversionPdo,
+                        BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                         TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
         productOrderEjb.publishProductOrderToSAP(conversionPdo, new MessageCollection(), true);
@@ -758,7 +758,7 @@ public class ProductOrderEjbTest {
         productPriceCache.refreshCache();
         Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                 .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                        10000,1000000,conversionPdo,
+                        BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                         TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
 
@@ -847,7 +847,7 @@ public class ProductOrderEjbTest {
 
         Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                 .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                        10000,1000000,conversionPdo,
+                        BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                         TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
 
@@ -995,7 +995,7 @@ public class ProductOrderEjbTest {
 
         Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                 .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                        10000,1000000,conversionPdo,
+                        BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                         TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
         productOrderEjb.publishProductOrderToSAP(conversionPdo, messageCollection, true);
@@ -1098,7 +1098,7 @@ public class ProductOrderEjbTest {
 
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString()))
                     .thenReturn(TestUtils.buildTestSapQuote(conversionPdo.getQuoteId(),
-                            10000,1000000,conversionPdo,
+                            BigDecimal.valueOf(10000), BigDecimal.valueOf(1000000),conversionPdo,
                             TestUtils.SapQuoteTestScenario.PRODUCTS_MATCH_QUOTE_ITEMS,companyCode.getSalesOrganization()));
 
             priceList.add(new QuotePriceItem(productOrderAddOn.getAddOn().getPrimaryPriceItem().getCategory(),
